Let the line number count be a long.
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/csv/trunk@1397898 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/csv/CSVLexer.java
@@ -192,7 +192,7 @@ class CSVLexer extends Lexer {
      */
     private Token encapsulatedTokenLexer(final Token tkn) throws IOException {
         // save current line number in case needed for IOE
-        final int startLineNumber = getLineNumber();
+        final long startLineNumber = getLineNumber();
         int c;
         while (true) {
             c = in.read();
